move calibrationdisplay to metrics/_plot
